refactor: #44 - 로딩인디케이터 뷰 분리
